backend/services/08-matching/requirements.txt
# FastAPI Opportunity Matching Service Dependencies - Task 4.5
# Core API framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.4.2

# Database connectivity
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Machine Learning and similarity
numpy==1.24.4
fuzzywuzzy==0.18.0
python-Levenshtein==0.25.0

# HTTP client for BGE service
httpx==0.25.2
aiohttp==3.9.1

# Configuration and environment
python-dotenv==1.0.0
pydantic-settings==2.0.3

# Logging and monitoring
structlog==23.2.0

# Caching (Redis integration)
redis==5.0.1
aioredis==2.0.1

# Background task processing
celery==5.3.4

# Testing (for development)
pytest==7.4.3
pytest-asyncio==0.21.1

# Development tools
black==23.11.0
isort==5.12.0
flake8==6.1.0